Cleanup
- Added dyvil.lang.Number.+()
- Added common interface for dyvil.lang.Boolean and
dyvil.lang.BooleanRef
- Implemented toString(), hashCode(), and equals(Object) for
dyvil.lang.Boolean and dyvil.lang.BooleanRef

+++ src/library/dyvil/lang/Predef.java
@@ -124,7 +124,7 @@ public class Predef
 	
 	public static @implicit int $hash$hash(long l)
 	{
-		return (int) (int) (l ^ (l >>> 32));
+		return (int) (l ^ l >>> 32);
 	}
 	
 	public static @implicit int $hash$hash(float f)
